continue using AC and AR registers
* make AC and AR 16-bit registers
* drop some unneeded type casts
* bring zc17 back, rework byte addressing
* make instructions use ar for memory addressing
* "microoperationalize" LS and RI
* minor formatting

+++ src/cpu/alu.c
@@ -98,7 +98,7 @@ void alu_16_add(int16_t reg, int16_t n, unsigned carry)
 		// V is never cleared
 	}
 
-	REG_RESTRICT_WRITE(IR_A, (uint16_t) ures);
+	REG_RESTRICT_WRITE(IR_A, ures);
 }
 
 
@@ -138,7 +138,7 @@ void alu_16_sub(int16_t reg, int16_t n)
 	} else {
 		// V is never cleared
 	}
-	REG_RESTRICT_WRITE(IR_A, (uint16_t) ures);
+	REG_RESTRICT_WRITE(IR_A, ures);
 }
 
 // -----------------------------------------------------------------------
